requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-modeltranslation==0.16.2
django-mptt==0.12.0
django-rosetta==0.9.6
idna==2.10
Pillow==8.1.2
polib==1.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
Unidecode==1.2.0
urllib3==1.26.4
